refactor(calendar): タグフィルタの TagFlatList を描画/行/DnD 基盤に分割する
TagFlatList.tsx（860行）から SortableTagItem / SortableParentBlock /
TagFlatListDnd（strategies + droppable）を切り出し。ロジックは書き換えずに
移動のみ。全ファイル 400 行未満。

Refs #1527
